+++ EggTimer/Model/Egg.swift
@@ -6,7 +6,7 @@ struct Egg: Identifiable {
     let name: String
     let imageName: String
     let eggType: EggType
-    let cookingTime: Int
+    let cookingTime: Double
 
     init(eggType: EggType) {
         self.eggType = eggType
@@ -15,15 +15,15 @@ struct Egg: Identifiable {
         case .soft:
             name = L10n.soft
             imageName = Asset.softEgg.name
-            cookingTime = 10
+            cookingTime = 10.0
         case .medium:
             name = L10n.medium
             imageName = Asset.mediumEgg.name
-            cookingTime = 20
+            cookingTime = 20.0
         case .hard:
             name = L10n.hard
             imageName = Asset.hardEgg.name
-            cookingTime = 30
+            cookingTime = 30.0
         }
     }
 }
